# api-docs.json
{"openapi":"3.0.1","info":{"title":"SBI Loan Management System — REST API","description":"**SBI Loan Management System** REST API\n\nDevSecOps Training Project — State Bank of India\n\n---\n\n### How to Authenticate\n\n1. Expand **Authentication** below → `POST /api/v1/auth/login`\n2. Click **Try it out** → fill in username and password → **Execute**\n3. Copy the `token` value from the response body\n4. Click **Authorize** (top right) → paste the token → **Authorize**\n\n---\n\n### Training Accounts\n\n| Username | Password | Role | Access Level |\n|---|---|---|---|\n| `admin@sbi.com` | `Admin@123` | ADMIN | Full access — all operations |\n| `officer@sbi.com` | `Officer@123` | USER | Read + own records only |\n\n---\n\n### Known Issues (Pre-Training — Fix During Labs)\n\n| # | Endpoint | Issue | Lab |\n|---|---|---|---|\n| 1 | `POST /api/v1/auth/login` | Credentials sent as query params — logged in URLs | Lab 1 |\n| 2 | `GET /api/v1/applications` | Returns raw entity — sensitive fields exposed to all users | Lab 1 |\n| 3 | `GET /api/v1/applications/search` | SQL injection via branch name parameter | Lab 2 |\n| 4 | All write endpoints | No `@PreAuthorize` — any user can create/update/delete | Lab 1 |\n\n---\n\n> ⚠️ **Swagger UI is disabled in production** (`springdoc.swagger-ui.enabled=false` in `application-prod.properties`). Swagger documents every endpoint and parameter — leaving it on in production gives attackers a complete API map.","contact":{"name":"SBI Technology Training Team","email":"training@sbi.co.in"},"license":{"name":"Internal Training Use Only","url":"https://www.sbi.co.in"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Local Development"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/applications/{id}/status":{"put":{"tags":["Loan Applications"],"summary":"Update application status","operationId":"updateStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanApplicationDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login and receive JWT token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/applications":{"get":{"tags":["Loan Applications"],"summary":"List all loan applications","operationId":"listAll_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanApplicationDTO"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Loan Applications"],"summary":"Create a new loan application","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoanApplicationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanApplicationDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/products":{"get":{"tags":["Loan Products"],"summary":"List all loan products","operationId":"listAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanProduct"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/products/{id}":{"get":{"tags":["Loan Products"],"summary":"Get loan product by ID","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanProduct"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/branches":{"get":{"tags":["Branches"],"summary":"List all branches","operationId":"listAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Branch"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/branches/{id}":{"get":{"tags":["Branches"],"summary":"Get branch by ID","operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Branch"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/applications/{id}":{"get":{"tags":["Loan Applications"],"summary":"Get application by ID","operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanApplicationDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/applications/search":{"get":{"tags":["Loan Applications"],"summary":"Search applications by branch name","operationId":"searchByBranch","parameters":[{"name":"branch","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"StatusUpdateRequest":{"required":["status"],"type":"object","properties":{"status":{"type":"string"}}},"LoanApplicationDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"applicantName":{"type":"string"},"email":{"type":"string"},"panNumber":{"type":"string"},"monthlyIncome":{"type":"number"},"cibilScore":{"type":"integer","format":"int32"},"loanAmountRequested":{"type":"number"},"status":{"type":"string","enum":["SUBMITTED","UNDER_REVIEW","APPROVED","REJECTED"]},"branchName":{"type":"string"},"productName":{"type":"string"}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"LoanApplicationRequest":{"required":["applicantName","branchId","cibilScore","email","loanAmountRequested","loanProductId","monthlyIncome","panNumber"],"type":"object","properties":{"applicantName":{"maxLength":100,"minLength":2,"pattern":"^[a-zA-Z\\s-]+$","type":"string"},"email":{"type":"string"},"panNumber":{"pattern":"^[A-Z]{5}[0-9]{4}[A-Z]{1}$","type":"string"},"monthlyIncome":{"type":"number"},"loanAmountRequested":{"type":"number"},"cibilScore":{"maximum":900,"minimum":300,"type":"integer","format":"int32"},"branchId":{"type":"integer","format":"int64"},"loanProductId":{"type":"integer","format":"int64"}}},"LoanProduct":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"productName":{"type":"string"},"minAmount":{"type":"number"},"maxAmount":{"type":"number"},"interestRate":{"type":"number"}}},"Branch":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"ifscCode":{"type":"string"},"city":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"Paste the token returned by POST /api/v1/auth/login. Do NOT include the word 'Bearer' here.","scheme":"bearer","bearerFormat":"JWT"}}}}